rename middleware to render and extract common rendering logic
