+++ packages/react-app-server/server/middlewares/index.ts
@@ -4,7 +4,7 @@ import serve from 'koa-static'
 
 import error from './error'
 import manifest from './manifest'
-import handleRender from './handler'
+import render from './render'
 
 const DIST_PUBLIC = '.dist/public'
 const DIST_STATIC = '.dist/static'
@@ -15,5 +15,5 @@ export default compose([
   serve(DIST_STATIC, { root: '/static' }),
   manifest(),
   error(),
-  handleRender,
+  render(),
 ])
